+++ src/components/tonieboxes/TonieboxCard.tsx
@@ -2,13 +2,14 @@ import React, { useEffect, useState } from 'react';
 import { Link } from 'react-router-dom';
 import { useTranslation } from 'react-i18next';
 import { Card, Button, Input, message, Modal, Badge } from 'antd';
-import { EditOutlined, SaveOutlined, CloseOutlined, SettingOutlined } from '@ant-design/icons';
+import { EditOutlined, SafetyCertificateOutlined, SaveOutlined, CloseOutlined, SettingOutlined } from '@ant-design/icons';
 import { defaultAPIConfig } from "../../config/defaultApiConfig";
 import { TeddyCloudApi } from "../../api";
 import { TonieboxModelSearch } from './TonieboxModelSearch';
 import { TonieboxSettingsPage } from './TonieboxSettingsPage';
 import { boxModelImages } from '../../util/boxModels';
 import { TonieCardProps } from '../../components/tonies/TonieCard';
+import { CertificateDragNDrop } from '../settings/CertificatesDragNDrop';
 
 const api = new TeddyCloudApi(defaultAPIConfig());
 
@@ -31,6 +32,7 @@ export const TonieboxCard: React.FC<{ tonieboxCard: TonieboxCardProps }> = ({ to
     const [tonieboxStatus, setTonieboxStatus] = useState<boolean>(false);
     const [lastPlayedTonieName, setLastPlayedTonieName] = useState<React.ReactNode>(null);
     const [isEditSettingsModalOpen, setIsEditSettingsModalOpen] = useState(false);
+    const [isUploadCertificatesModalOpen, setIsUploadCertificatesModalOpen] = useState(false);
     const [isModelModalOpen, setIsModelModalOpen] = useState(false);
     const [activeModel, setActiveModel] = useState<string>(tonieboxCard.boxModel);
     const [selectedModel, setSelectedModel] = useState<string>(tonieboxCard.boxModel);
@@ -68,7 +70,7 @@ export const TonieboxCard: React.FC<{ tonieboxCard: TonieboxCardProps }> = ({ to
     const selectBoxImage = (id: string) => {
         const selectedImage = boxModelImages.find(item => item.id === id);
         if (selectedImage) {
-            setBoxImage(<img src={selectedImage.img_src} alt="" />);
+            setBoxImage(<img src={selectedImage.img_src} alt="" style={{ }}/>);
         }
     };
 
@@ -83,6 +85,20 @@ export const TonieboxCard: React.FC<{ tonieboxCard: TonieboxCardProps }> = ({ to
             </>
         );
     }
+
+    // certificates
+    const handleUploadCertificatesClick = () => {
+        showUploadCertificatesModal();
+    };
+    const showUploadCertificatesModal = () => {
+        setIsUploadCertificatesModalOpen(true);
+    };
+    const handleUploadCertificatesOk = async () => {
+        setIsUploadCertificatesModalOpen(false);
+    };
+    const handleUploadCertificatesCancel = () => {
+        setIsUploadCertificatesModalOpen(false);
+    };
 
     // Settings
     const handleEditSettingsClick = () => {
@@ -199,16 +215,24 @@ export const TonieboxCard: React.FC<{ tonieboxCard: TonieboxCardProps }> = ({ to
                 title={<span><Badge dot status={tonieboxStatus ? "success" : "error"} /> {tonieboxName}</span>}
                 cover={<div style={{ position: 'relative' }}>
                     {lastPlayedTonieName}
-                    <img src={boxImage.props.src} alt="Box" style={{ width: '100%', height: 'auto' }} />
+                    <img src={boxImage.props.src} alt="Box" style={{...boxImage.props.style, width: '100%', height: 'auto' }} />
                 </div>}
-                actions={[<span key="settings" onClick={handleEditSettingsClick} >
-                    <SettingOutlined key="edit" style={{ marginRight: 8 }} />{t("tonieboxes.editTonieboxSettingsModal.editTonieboxSettingsLabel")}
-                </span>]}
+                actions={[
+                    	<span key="settings" onClick={handleUploadCertificatesClick} >
+                            <SafetyCertificateOutlined key="certificate" style={{ marginRight: 8 }} />{t("tonieboxes.uploadTonieboxCertificatesModal.Label")}
+                        </span>, 
+                        <span key="settings" onClick={handleEditSettingsClick} >
+                            <SettingOutlined key="edit" style={{ marginRight: 8 }} />{t("tonieboxes.editTonieboxSettingsModal.editTonieboxSettingsLabel")}
+                        </span>
+                ]}
             >
                 <Meta description={<div>{"MAC: " + tonieboxCard.ID.replace(/(.{2})(?=.)/g, "$1:")}</div>} />
             </Card >
             <Modal title={t("tonieboxes.editTonieboxSettingsModal.editTonieboxSettings", { "name": tonieboxCard.boxName })} width='auto' open={isEditSettingsModalOpen} onOk={handleEditSettingsOk} onCancel={handleEditSettingsCancel}>
                 <TonieboxSettingsPage overlay={tonieboxCard.ID} />
+            </Modal>
+            <Modal title={t("tonieboxes.uploadTonieboxCertificatesModal.uploadTonieboxCertificates", { "name": tonieboxCard.boxName })} width='auto' open={isUploadCertificatesModalOpen} onOk={handleUploadCertificatesOk} onCancel={handleUploadCertificatesCancel}>
+                <CertificateDragNDrop overlay={tonieboxCard.ID} />
             </Modal>
             <Modal title={t("tonieboxes.editModelModal.editModel", { "name": tonieboxCard.boxName })} open={isModelModalOpen} onOk={handleModelOk} onCancel={handleModelCancel} afterClose={() => setSearchFieldValue('')}>
                 <p><Input name="boxName" value={boxName} onChange={(e) => setBoxName(e.target.value)}
